+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Schematic:
   - Support for SHEETNAME, SHEETFILE and SHEETPATH in the title block
   - Support for the `jumper_pin_groups` attribute of symbols. (#942)
+  - Some support for properties defined at sheet level. (#941)
 - Globals:
   - `schematic_sheet_name_workaround` to workaround SHEETNAME on page 1
     (See #933)